flask_dir/requirements.txt
attrs==20.2.0
boto==2.49.0
boto3==1.16.1
botocore==1.19.1
click==7.1.2
Flask==1.1.2
importlib-metadata==2.0.0
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
mysql-connector-python==8.0.21
packaging==20.4
pluggy==0.13.1
protobuf==3.13.0
py==1.9.0
pyparsing==2.4.7
pytest==6.1.1
python-dateutil==2.8.1
python-dotenv==0.14.0
s3transfer==0.3.3
six==1.15.0
toml==0.10.1
urllib3==1.25.11
Werkzeug==1.0.1
zipp==3.3.0